Organize some imports and configs
