--- app/src/config/api.js
@@ -1,1 +1,0 @@
-export const BASE_API_URL = process.env.REACT_APP_API_BASE_URL || '/api';
