+++ app/src/config/index.js
@@ -1,5 +1,3 @@
-export * from './api';
-export * from './colors';
 export * from './competition';
 export * from './levels';
 export * from './player';
